requirements.txt
# ── Núcleo ──────────────────────────────────
python-dotenv>=1.0.1
pydantic>=2.7.0
httpx>=0.27.0
tenacity>=8.3.0
loguru>=0.7.2

# ── STT: Whisper via API OpenAI ──────────────
openai>=1.30.1

# ── Audio I/O ────────────────────────────────
sounddevice>=0.4.6
numpy>=2.0.0
pygame>=2.5.2

# ── Automação Web ────────────────────────────
playwright>=1.44.0

# ── Automação de SO ──────────────────────────
pyautogui>=0.9.54
Pillow>=10.3.0
pygetwindow>=0.0.9
keyboard>=0.13.5

# ── Windows específico ────────────────────────
pywin32>=306

# ── Brain / RAG local ────────────────────────
rank-bm25>=0.2.2
